+++ src/App.vue
@@ -445,7 +445,6 @@ onMounted(() => {
       :analysis-tables="viewer.acqImageCollectionDocument.value.data.analysis_tables"
       :preferred-table="preferredAnalysisTableForUrl(viewer.acqImageCollectionDocument.value.url)"
       :load-table="viewer.loadCollectionTable"
-      :load-json="viewer.loadCollectionJson"
       :selected-acq-image-id="viewer.selectedAcqImageId.value"
       :selected-channel="viewer.selectedChannel.value"
       :selected-roi-id="viewer.selectedRoiId.value"